requirements.txt
tensorflow==2.2.0
h5py==2.10.0
imgaug==0.2.9
numpy==1.18.4
matplotlib==3.2.1
scipy==1.4.1
scikit_image==0.16.2
streamlit==0.61.0
Keras==2.2.3
opencv_contrib_python==4.1.2.30
ipython==7.15.0
Pillow==7.1.2
skimage==0.0
